+++ cmd/hidden_lake/messenger/pkg/app/app.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
-	"sync"
 
 	"github.com/number571/go-peer/cmd/hidden_lake/messenger/internal/config"
 	"github.com/number571/go-peer/cmd/hidden_lake/messenger/internal/database"
@@ -25,7 +24,6 @@ var (
 
 type sApp struct {
 	fState state.IState
-	fMutex sync.Mutex
 
 	fConfig config.IConfig
 	fPathTo string
@@ -118,9 +116,6 @@ func (p *sApp) Run(pCtx context.Context) error {
 }
 
 func (p *sApp) stop() error {
-	p.fMutex.Lock()
-	defer p.fMutex.Unlock()
-
 	err := interrupt.CloseAll([]types.ICloser{
 		p.fIntServiceHTTP,
 		p.fIncServiceHTTP,
@@ -130,6 +125,5 @@ func (p *sApp) stop() error {
 	if err != nil {
 		return fmt.Errorf("close/stop all: %w", err)
 	}
-
 	return nil
 }